Updated to discord.js v12, added logging channel
The bot now works with discord.js v12, and instructions that use changed methods have been updated accordingly.
More logging lines have been added to log the different kind of commands users can use,
and for each log to the console a message is now also sent to a dedicated channel in the bot's server,
so there's a more permanent logging solution.